alarms: Skip nested bitfield paths
These generate non-sensical paths because of how checkAlarms is implemented.
checkAlarms (the version we're using) automatically derives type from the path,
HOWEVER: it'll go into parent nodes, until it finds the typename. This results
in nested status nodes having the same typename as one of the parent nodes,
which means that all of its alarms got duplicated.